logI-job代码优化：增加logi相关的 rest 接口

+++ logi-job-spring-boot-starter/src/main/java/com/didiglobal/logi/job/core/task/TaskLockServiceImpl.java
@@ -101,6 +101,9 @@ public class TaskLockServiceImpl implements TaskLockService {
             logITaskLockPO.getCreateTime().getTime() / 1000 + logITaskLockPO.getExpireTime() < current)
             .map( LogITaskLockPO::getId)
             .collect(Collectors.toList());
+
+    if(CollectionUtils.isEmpty(taskLockIdList)){return true;}
+
     int result = logITaskLockMapper.deleteByIds(taskLockIdList);
     return result > 0 ? true : false;
   }
